Refactor setUpdateAverageGrade method: change parameter type from int to float

+++ Prog 01_Ozuna_Salazar_Jorge_Alejandro/include/Student.h
@@ -14,9 +14,9 @@ class Student
         ~Student();
 
         void showStudentData();
-        void setUpdateAverageGrade(int);
+        void setUpdateAverageGrade(float);
         void setUpdateName(std::string);
-        
+
         bool doesTheStudenPass();
         bool isTheStudenoutstanding();
 };
